Fixed bug associated with issue 1.
The covariance matrix of the algebraic parameters associated with
the original (unnormalised) coordinate system was incorrect due
to an oversight in the implementation. The issue did not
affect the covariance matrix of the algebraic parameters in
the normalised coordinate system, nor did it affect the covariance
matrix of the geometric parameters.

The problem has been now been resolved.

I also fixed some typos in run_example.m